animation working but lots to do
