+++ clean.sh
