+++ controller/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rayscale",
-  "version": "1.5.0",
+  "version": "2.0.0",
   "description": "Container based distributed raytracer - controller",
   "main": "dist/server.js",
   "scripts": {
