+++ controller/src/lib/job-input.ts
@@ -1,6 +1,6 @@
 //
 // Unparsed Job as received from the API (as YAML)
-// ------------------------------------------------------
+// -----------------------------------------------
 // Ben C, May 2018
 //
