+++ controller/src/lib/job-stats.ts
@@ -1,3 +1,9 @@
+//
+// Just some stats and counters
+// ----------------------------
+// Ben C, May 2018
+//
+
 export class JobStats {
   raysCreated: number;
   raysCast: number;
