+++ controller/src/lib/tracer.ts
@@ -1,6 +1,6 @@
 //
 // Basic details of a remote tracer
-// ------------------------------------------------------------------
+// --------------------------------
 // Ben C, May 2018
 //
